% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPI_SPEI_at_specific_day.R
\name{daily_spei}
\alias{daily_spei}
\title{Standardised precipitation evapotranspiration index calculated for a specific day. 
 This function is a wrapper function of daily.spi}
\usage{
daily_spei(params_list, P_lyr)
}
\arguments{
\item{params_list}{A list object that contains a 'SpatRaster' of 365 layers for each parameter of the selected distribution.}

\item{P_lyr}{'SpatRaster' object that contains spatially-distributed monthly data for a specific date (e.g., "%Y-02-28).
This object contains the daily accumulated values of a specific day according to the scale provided. This object can be computed with
the 'spi_agregate_daily' function.
This 'SpatRaster' must only contain the days that corresponds to the specific selection.}
}
\value{
This function returns one layer of SPI-n according to a selected day. The input object 'Prod_data'
   has to be created before, which can be achieved with the 'spi.agregate_daily' function.
}
\description{
Standardised precipitation evapotranspiration index calculated for a specific day. 
 This function is a wrapper function of daily.spi
}
