% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meteorological_drought_utils.R
\name{.calculate_pze}
\alias{.calculate_pze}
\title{Calculate the probability of zero}
\usage{
.calculate_pze(x, ref_period, dates)
}
\arguments{
\item{x}{Numerical vector.}

\item{ref_period}{Period of reference c(ref_start, ref_end) with the format "\%Y-\%m-}

\item{dates}{Character vector with the dates of x.}
}
\description{
Calculate the probability of zero
}
