% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combined_drought_indices_utils.R
\name{.cdi_vals}
\alias{.cdi_vals}
\title{Utils function to calculate the values to be used in the CDI}
\usage{
.cdi_vals(zSPI, zSM, zfAPAR)
}
\arguments{
\item{zSPI}{'SpatRaster' object that contains the zSPI data that will be used to calculate the CDI 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.}

\item{zSM}{'SpatRaster' object that contains the zSM data that will be used to calculate the CDI 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.}

\item{zfAPAR}{'SpatRaster' object that contains the zSfAPAR data that will be used to calculate the CDI 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.}
}
\value{
Return the 3-digit values to compute the CDI.
}
\description{
Utils function to calculate the values to be used in the CDI
}
