% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meteorological_drought_utils.R
\name{.deciles}
\alias{.deciles}
\title{Utils function to calculate deciles}
\usage{
.deciles(x)
}
\arguments{
\item{x}{Numerical vector.}
}
\value{
Numerical vector with the corresponding deciles
}
\description{
Utils function to calculate deciles
}
