% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agricultural_drought_utils.R
\name{.essmi}
\alias{.essmi}
\title{Utils function to calculate the Empirical Standardised Soil Moisture Index (ESSMI)}
\usage{
.essmi(
  x,
  dates,
  scale,
  ref_start,
  ref_end,
  distribution,
  bw,
  missing_ratio,
  ...
)
}
\arguments{
\item{x}{Numerical vector.}

\item{dates}{Vector of dates that is extracted from the 'SpatRaster' in the 'spatial_essmi' function.}

\item{scale}{Integer value that represents the time scale at which the ESSMI will be computed.}

\item{ref_start}{optional value that represents the starting point of the reference period used for computing the index. 
The date should be introduced as '\%Y-\%m'. For example: "1989-02".
The default is NULL, which indicates that the first layer in the 'SpatRaster' will be used as starting point.}

\item{ref_end}{Optional value that represents the ending point of the reference period used for computing the index. 
The date should be introduced as '\%Y-\%m'. For example: "1989-02".
The default is NULL, which indicates that the last layer in the 'SpatRaster' will be used as ending point.}

\item{distribution}{Optional value indicating the name of the distribution function to be used in the Kernel Density Estimation
(one of 'Gaussian', 'Rectangular', 'Triangular', 'Epanechnikov', 'Biweight', 'Cosine' and 'Optcosine'). Defaults to 'Gaussian' for ESSMI.}

\item{bw}{the smoothing bandwidth to be used. The kernels are scaled such that this is the standard deviation of the smoothing kernel. 
The default is set to the methods of Sheather & Jones (1991) to select the bandwidth using pilot estimation of derivatives.
Please see 'Bandwidth Selectors for Kernel Density Estimation' from 'stats' for more information.}

\item{missing_ratio}{Ratio of missing data that is acceptable for the computation. Set to 0.2 by default (20\%).}

\item{...}{Additional variables that can be used for the 'density' function.}
}
\value{
Numerical vector with the corresponding to the ESSMI
}
\description{
Utils function to calculate the Empirical Standardised Soil Moisture Index (ESSMI)
}
