% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combined_drought_indices_utils.R
\name{.get_cdi}
\alias{.get_cdi}
\title{Calculate the CDI index}
\usage{
.get_cdi(x)
}
\arguments{
\item{x}{vector object with the three-digit CDI values.}
}
\value{
Return a vector with CDI values.
}
\description{
Calculate the CDI index
}
