% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climdex_utils.R
\name{.idx.cdd}
\alias{.idx.cdd}
\title{CDD}
\usage{
.idx.cdd(rst)
}
\arguments{
\item{rst}{'SpatRaster' object with daily precipitation values.}
}
\value{
Maximum length of dry spell: maximum number of consecutive days with RR < 1mm
}
\description{
CDD
}
