% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climdex_utils.R
\name{.idx.cwd}
\alias{.idx.cwd}
\title{CWD}
\usage{
.idx.cwd(rst)
}
\arguments{
\item{rst}{'SpatRaster' object with daily precipitation values.}
}
\value{
Maximum length of wet spell: maximum number of consecutive days with RR >= 1mm
}
\description{
CWD
}
