% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climdex_utils.R
\name{.idx.prcptot}
\alias{.idx.prcptot}
\title{PRCPTOT}
\usage{
.idx.prcptot(rst)
}
\arguments{
\item{rst}{'SpatRaster' object with daily precipitation values.}
}
\value{
Annual total precipitation on wet days
}
\description{
PRCPTOT
}
