% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climdex_utils.R
\name{.idx.rnnmm}
\alias{.idx.rnnmm}
\title{Rnnmm (threshold should be applied: for R10mm thres = 10; for R20mm thres = 20;
for Rnnmm thres = nn)}
\usage{
.idx.rnnmm(rst, thres)
}
\arguments{
\item{rst}{'SpatRaster' object with daily precipitation values.}

\item{thres}{Threshold value that corresponds to nn.}
}
\value{
Annual count of days when PRCP >= a designed threshold
}
\description{
Rnnmm (threshold should be applied: for R10mm thres = 10; for R20mm thres = 20;
for Rnnmm thres = nn)
}
