% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climdex_utils.R
\name{.idx.rnnp}
\alias{.idx.rnnp}
\title{Rnnp (threshold should be applied: for R95p thres = 0.95; for R99p thres = 0.99)}
\usage{
.idx.rnnp(rst, thres)
}
\arguments{
\item{rst}{'SpatRaster' object with daily precipitation values.}
}
\value{
Annual total PRCP when RR > a designed percentile
}
\description{
Rnnp (threshold should be applied: for R95p thres = 0.95; for R99p thres = 0.99)
}
