% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climdex_utils.R
\name{.idx.rx1day}
\alias{.idx.rx1day}
\title{RX1day}
\usage{
.idx.rx1day(rst)
}
\arguments{
\item{rst}{'SpatRaster' object with daily precipitation values.}
}
\value{
Maximum 1-day precipitation
}
\description{
RX1day
}
