% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climdex_utils.R
\name{.idx.rx5day}
\alias{.idx.rx5day}
\title{RX5day}
\usage{
.idx.rx5day(rst)
}
\arguments{
\item{rst}{'SpatRaster' object with daily precipitation values.}
}
\value{
Maximum 5-day precipitation
}
\description{
RX5day
}
