% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climdex_utils.R
\name{.idx.sdii}
\alias{.idx.sdii}
\title{SDII}
\usage{
.idx.sdii(rst)
}
\arguments{
\item{rst}{'SpatRaster' object with daily precipitation values.}
}
\value{
Simple precipitation intensity index
}
\description{
SDII
}
