% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPI_SPEI_at_specific_day_utils.R
\name{.kalman_filter}
\alias{.kalman_filter}
\title{Utils function to apply Kalman filter}
\usage{
.kalman_filter(x, H, ...)
}
\arguments{
\item{x}{Numerical vector.}

\item{H}{Covariance matrix or array of disturbance terms \epsilon_tϵ of observation equations.}

\item{...}{}
}
\description{
Utils function to apply Kalman filter
}
