% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trend_analyses.R
\name{.MKpval}
\alias{.MKpval}
\title{Utils non-parametric man-Kendall trend test}
\usage{
.MKpval(x)
}
\arguments{
\item{x}{numerical vector.}
}
\value{
returns a numerical vector representing the p-value, Kendall_Score-S,
  Variance_of_Kendall_Score-varS, and Tau.
}
\description{
Utils non-parametric man-Kendall trend test
}
