% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meteorological_drought_utils.R
\name{.pni}
\alias{.pni}
\title{Utils function to calculate percent of normal}
\usage{
.pni(x, dates)
}
\arguments{
\item{x}{Numerical vector.}

\item{dates}{Vector of dates that is extracted from the 'SpatRaster' in the 'spatial_spei' function.}
}
\value{
Numerical vector with the corresponding percent of normal precipitation
}
\description{
Utils function to calculate percent of normal
}
