% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPI_SPEI_at_specific_day_utils.R
\name{.spei_daily.spei}
\alias{.spei_daily.spei}
\title{Utils function to compute the SPI or SPEI using a vector. This function is based on the code presented in the SPEI package
 from 'sbegueria' https://github.com/sbegueria/SPEI/blob/master/R/spei.R}
\usage{
.spei_daily.spei(params_matrix, P_matrix, pze_matrix, trgt, distribution)
}
\arguments{
\item{params_matrix}{matrix obtained from the parameters list}

\item{P_matrix}{matrix of the required P layer}

\item{pze_matrix}{matrix of the probability of zero layer}

\item{trgt}{The day for which the function will be computed. The default is NULL, indicating that the last day of the data
will be selected.}

\item{distribution}{Character value indicating the name of the distribution function to be used for computing the SPI 
(one of 'log-Logistic', 'Gamma' and 'PearsonIII'). Defaults to 'log-Logistic' for SPEI.}
}
\description{
Utils function to compute the SPI or SPEI using a vector. This function is based on the code presented in the SPEI package
 from 'sbegueria' https://github.com/sbegueria/SPEI/blob/master/R/spei.R
}
