% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combined_drought_indices_utils.R
\name{.spi_bool}
\alias{.spi_bool}
\title{Utils function to calculate the boolean layers to calculate the zSPI}
\usage{
.spi_bool(SPI_data, threshold)
}
\arguments{
\item{SPI_data}{'SpatRaster' object that contains the SPI data that will be used to calculate the zSPI. 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.}

\item{threshold}{Threshold to construct the Boolean layers.}
}
\description{
Utils function to calculate the boolean layers to calculate the zSPI
}
