% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agricultural_drought_utils.R
\name{.tci}
\alias{.tci}
\title{Utils function to calculate the Temperature Condition Index}
\usage{
.tci(x, ...)
}
\arguments{
\item{x}{Numerical vector.}
}
\value{
Numerical vector with the corresponding to the TCI
}
\description{
Utils function to calculate the Temperature Condition Index
}
