% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agricultural_drought_utils.R
\name{.vci}
\alias{.vci}
\title{Utils function to calculate the Vegetation Condition Index}
\usage{
.vci(x, ...)
}
\arguments{
\item{x}{Numerical vector.}
}
\value{
Numerical vector with the corresponding to the VCI
}
\description{
Utils function to calculate the Vegetation Condition Index
}
