% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agricultural_drought_utils.R
\name{.zscore}
\alias{.zscore}
\title{Utils function to calculate the z-score}
\usage{
.zscore(x, dates)
}
\arguments{
\item{x}{Numerical vector.}

\item{dates}{Vector of dates that is extracted from the 'SpatRaster' in the 'spatial_zscore' function.}
}
\value{
Numerical vector with the corresponding to the zFAPAR
}
\description{
Utils function to calculate the z-score
}
