% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monthly_data_utils.R
\name{import_monthly_data}
\alias{import_monthly_data}
\title{Import a monthly SpatRaster from monthly NCDF or GeoTIFF files.}
\usage{
import_monthly_data(rst_path, start_date, end_date, years = NULL)
}
\arguments{
\item{rst_path}{File path to either:
\itemize{
 \item'GTiff files': Monthly raster files that include the date in the file names (e.g., 'ProductX_1989-02.tif'). The dates must have the format '\%Y-\%m'.
 \item'NCDF files': Annual raster files that contain the monthly layers respective for that year. The file name must include the date (e.g., 'ProductX_1989.nc'). The dates must have the format '\%Y'.
}}

\item{start_date}{Position where the dates of the raster files start. For example, start_date = 10 for 'ProductX_1989-02.tif' as the date start in the 10th character.}

\item{end_date}{Position where the dates of the raster files end. For example, end_date = 16 for 'ProductX_1989-02.tif', and end_date = 13 for 'ProductX_1989.nc'.}

\item{years}{Integer vector indicating the years that should be imported. For example, years <- 2000:2010 for importing monthly data from 2000 to 2010.}
}
\value{
SpatRaster of the monthly data that will be used in further analysis.
}
\description{
Import a monthly SpatRaster from monthly NCDF or GeoTIFF files.
}
