% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trend_analyses.R
\name{mk_spatial}
\alias{mk_spatial}
\title{Non-parametric Man-Kendall trend test}
\usage{
mk_spatial(rst, vct = NULL, conf_level = 0.95, p_thres = 0.05)
}
\arguments{
\item{rst}{'SpatRaster' object.}

\item{vct}{Vector file of the study area (Optional). It will be used to crop 
the spatial extent of the raster files if required. This parameter is set 
to 'NULL'}

\item{conf_level}{Level of significance. Set to 0.95 by default.}

\item{p_thres}{threshold of p-value. Set to 0.05 by default.}
}
\value{
returns gridded layers(i.e., grid-cells where p-value < p_thres, 
 p-value, Kendall_Score-S, Variance_of_Kendall_Score-varS, Tau values, 
 "Sen's_Slope", "Total_change") and a vector layer (i.e., grid-cells where p-value < p_thres).
}
\description{
Non-parametric Man-Kendall trend test
}
