% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meteorological_drought.R
\name{params_spi}
\alias{params_spi}
\title{Calculation of the parameters of a selected probability distribution for a selected day (SPI).}
\usage{
params_spi(
  Prod_data,
  trgt = NULL,
  ref_start = NULL,
  ref_end = NULL,
  distribution = "Gamma",
  fit = "ub-pwm",
  package = "SCI"
)
}
\arguments{
\item{Prod_data}{'SpatRaster' object that contains spatially-distributed monthly data for a specific date (e.g., "%Y-02-28).
This object contains the daily accumulated values of a specific day according to the scale provided. This object can be computed with
the 'spi_agregate_daily' function.
This 'SpatRaster' must only contain the days that corresponds to the specific selection.}

\item{trgt}{The day for which the function will be computed. The default is NULL, indicating that the last day of the data
will be used to return the SPI values.}

\item{ref_start}{optional value that represents the starting point of the reference period used for computing the index. 
The date should be introduced as '\%Y-\%m'. For example: "1989-02".
The default is NULL, which indicates that the first layer in the 'SpatRaster' will be used as starting point.}

\item{ref_end}{Optional value that represents the ending point of the reference period used for computing the index. 
The date should be introduced as '\%Y-\%m'. For example: "1989-02".
The default is NULL, which indicates that the last layer in the 'SpatRaster' will be used as ending point.}

\item{distribution}{Optional value indicating the name of the distribution function to be used for computing the SPI 
(one of 'log-Logistic', 'Gamma' and 'PearsonIII'). Defaults to 'Gamma' for SPI.}

\item{fit}{Optional value indicating the name of the method used for computing the distribution function parameters 
(one of 'ub-pwm', 'pp-pwm' and 'max-lik'). Defaults to 'ub-pwm'.}

\item{package}{Either 'SCI' or 'SPEI'. Should the SCI or SPEI package be used in the implementation?}
}
\value{
This function returns the fit parameters of the selected distribution according to a selected day. The input object 'Prod_data'
   has to be created before, which can be achieved with the 'spi.agregate_daily' function.
}
\description{
Calculation of the parameters of a selected probability distribution for a selected day (SPI).
}
