% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climdex.R
\name{R99pTOT}
\alias{R99pTOT}
\title{Contribution to total precipitation from very wet days (99p)}
\usage{
R99pTOT(rst_path, vct = NULL, start_date, end_date, init_month = NULL)
}
\arguments{
\item{rst_path}{File path to either:
\itemize{
 \item'GTiff files': Daily raster files that include the date in the file names (e.g., 'ProductX_1989-02-28.tif'). The dates must have the format  '\%Y-\%m-\%d'. 
 \item'NCDF files': Monthly raster files that contain the daily layers respective for that month. The file name must include the date (e.g., 'ProductX_1989-02.nc'). The dates must have the format  '\%Y-\%m'. 
}}

\item{vct}{Vector file of the study area (Optional). It will be used to crop 
the spatial extent of the raster files if required. This parameter is set 
to 'NULL'.}

\item{start_date}{Position where the dates of the raster files start. For example, start_date = 10 for 'ProductX_1989-02-28.tif' as the date start in the 10th character.}

\item{end_date}{Position where the dates of the raster files end. For example, end_date = 19 for 'ProductX_1989-02-28.tif', and end_date = 16 for 'ProductX_1989-02.nc'.}

\item{init_month}{Numeric value that represents the initial month that will be used in the calculation of the indices for each year (1 = Jan, 2 = Feb, ...., 12 = December). 
This parameter is set to NULL meaning that if not specified, the year will start in January.}
}
\value{
raster layers with the spatial distribution of the contribution to total precipitation from very wet days (99p) for each year of analysis.
}
\description{
Contribution to total precipitation from very wet days (99p)
}
