% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPI_SPEI_at_specific_day.R
\name{read_parameters}
\alias{read_parameters}
\title{Import the params_list object to a specific directory}
\usage{
read_parameters(dir, folder_name = NULL)
}
\arguments{
\item{dir}{Path to the directory where the data is stored}

\item{folder_name}{Which is the name of the parent folder that will be created to store the results? Defaults to NULL (name of folder = Results)}
}
\value{
imports the params_list object that is saved to a specific location
}
\description{
Import the params_list object to a specific directory
}
