% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combined_drought_indices.R
\name{spatial_cdi}
\alias{spatial_cdi}
\title{Combined Drought Indicator (CDI; Cammalleri et al. 2021)}
\usage{
spatial_cdi(
  SPI1_data,
  SPI3_data,
  zSM,
  zfAPAR,
  threshold_spi1 = -2,
  threshold_spi3 = -1
)
}
\arguments{
\item{SPI1_data}{'SpatRaster' object that contains spatially-distributed SPI-1 data that will be used to calculate the CDI. 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.}

\item{SPI3_data}{'SpatRaster' object that contains spatially-distributed SPI-3 data that will be used to calculate the CDI. 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.}

\item{zSM}{'SpatRaster' object that contains spatially-distributed zSM data that will be used to calculate the CDI (to calculate this index see the 'spatial_zscore' function). 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.}

\item{zfAPAR}{'SpatRaster' object that contains spatially-distributed zfAPAR data that will be used to calculate the CDI (to calculate this index see the 'spatial_zscore' function). 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.
@param threshold_spi1 Threshold to convert the SPI-1 data to a Boolean 'SpatRaster' (1 values > threshold; 0, otherwise). The default value of this threshold is -2 (Cammalleri et al. 2021).
#'  @param threshold_spi3 Threshold to convert the SPI-3 data to a Boolean 'SpatRaster' (1 values > threshold; 0, otherwise). The default value of this threshold is -1 (Cammalleri et al. 2021).}
}
\value{
Spatially-distributed CDI values.
}
\description{
Combined Drought Indicator (CDI; Cammalleri et al. 2021)
}
