% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meteorological_drought.R
\name{spatial_deciles}
\alias{spatial_deciles}
\title{Function to calculate the Deciles method}
\usage{
spatial_deciles(P_data)
}
\arguments{
\item{P_data}{'SpatRaster' object that contains spatially-distributed monthly precipitation data that will be used to calculate the SPI. 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.}
}
\value{
Spatially-distributed Deciles values.
}
\description{
Function to calculate the Deciles method
}
