% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meteorological_drought.R
\name{spatial_pni}
\alias{spatial_pni}
\title{Function to calculate the Percent of Normal Index}
\usage{
spatial_pni(P_data)
}
\arguments{
\item{P_data}{'SpatRaster' object that contains spatially-distributed monthly precipitation data that will be used to calculate the 
Percent of Normal Index. This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can 
be set with the function time of the terra package.}
}
\value{
Spatially-distributed Percent of Normal Index.
}
\description{
Function to calculate the Percent of Normal Index
}
