% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meteorological_drought.R
\name{spatial_spei}
\alias{spatial_spei}
\title{Wrapper function from the SPEI package to calculate the spatially-distributed Standardized Precipitation-Evapotranspiration Index (SPEI)}
\usage{
spatial_spei(
  P_data,
  PE_data,
  scale,
  ref_start = NULL,
  ref_end = NULL,
  distribution = "log-Logistic",
  fit = "ub-pwm",
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{P_data}{'SpatRaster' object that contains spatially-distributed monthly precipitation data that will be used to calculate the SPEI. 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.}

\item{PE_data}{'SpatRaster' object that contains spatially-distributed monthly potential evaporation (also referred to as potential evapotranspiation) 
data that will be used to calculate the SPEI. This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. 
They can be set with the function time of the terra package.}

\item{scale}{Integer value that represents the time scale at which the SPI will be computed.}

\item{ref_start}{Optional value that represents the starting point of the reference period used for computing the index. 
The date should be introduced as '\%Y-\%m'. For example: "1989-02".
The default is NULL, which indicates that the first layer in the 'SpatRaster' will be used as starting point.}

\item{ref_end}{Optional value that represents the ending point of the reference period used for computing the index. 
The date should be introduced as '\%Y-\%m'. For example: "1989-02".
The default is NULL, which indicates that the last layer in the 'SpatRaster' will be used as ending point.}

\item{distribution}{Optional value indicating the name of the distribution function to be used for computing the SPEI 
(one of 'log-Logistic', 'Gamma' and 'PearsonIII'). Defaults to 'log-Logistic' for SPEI.}

\item{fit}{Optional value indicating the name of the method used for computing the distribution function parameters 
(one of 'ub-pwm', 'pp-pwm' and 'max-lik'). Defaults to 'ub-pwm'.}

\item{na.rm}{Should the NA values be removed? Set to TRUE.}

\item{...}{Additional variables that can be used for the 'spi' function of the SPEI package.}
}
\value{
Spatially-distributed SPEI values.
}
\description{
Wrapper function from the SPEI package to calculate the spatially-distributed Standardized Precipitation-Evapotranspiration Index (SPEI)
}
