% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrological_drought.R
\name{spatial_swei}
\alias{spatial_swei}
\title{Snow Water Equivalent Index (SWEI)}
\usage{
spatial_swei(
  SWE_data,
  scale,
  start_month = NULL,
  end_month = NULL,
  missing_ratio = 0.2,
  ...
)
}
\arguments{
\item{scale}{Integer value that represents the time scale at which the ESSMI will be computed.}

\item{start_month}{Optional value that represents the starting point of the period that will be returned from the Index.
The idea is not to return periods where SWE is not present. 
The date should be introduced as '\%Y-\%m'. For example: "1989-02".
The default is NULL, which indicates that the first layer in the 'SpatRaster' will be used as starting point.}

\item{end_month}{Optional value that represents the ending point of the period that will be returned from the Index.
The idea is not to return periods where SWE is not present. 
The date should be introduced as '\%Y-\%m'. For example: "1989-02".
The default is NULL, which indicates that the last layer in the 'SpatRaster' will be used as ending point.}

\item{missing_ratio}{Ratio of missing data that is acceptable for the computation. Set to 0.2 by default (20\%).}

\item{...}{Additional variables that can be used for the 'density' function.}

\item{SM_data}{'SpatRaster' object that contains spatially-distributed monthly soil moisture data that will be used to calculate the ESSMI. 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.}
}
\value{
Spatially-distributed SWEI values.
}
\description{
Snow Water Equivalent Index (SWEI)
}
