% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agricultural_drought.R
\name{spatial_tci}
\alias{spatial_tci}
\title{Temperature Condition Index (TCI; Kogan 1995)}
\usage{
spatial_tci(BT_data, per_period = TRUE)
}
\arguments{
\item{BT_data}{'SpatRaster' object that contains spatially-distributed Brightness Temperature (BT) data that will be used to calculate the TCI. 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.}

\item{per_period}{Should the results be calculated considering each period separately? Set to TRUE.}
}
\value{
Spatially-distributed TCI values.
}
\description{
Temperature Condition Index (TCI; Kogan 1995)
}
