% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agricultural_drought.R
\name{spatial_vhi}
\alias{spatial_vhi}
\title{Vegetation Health Index (VHI; Kogan 1995)}
\usage{
spatial_vhi(VCI_data, TCI_data, alpha = 0.5, resampling2low = TRUE)
}
\arguments{
\item{VCI_data}{'SpatRaster' object that contains spatially-distributed VCI data that will be used to calculate the VHI. 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.}

\item{TCI_data}{'SpatRaster' object that contains spatially-distributed TCI data that will be used to calculate the VHI. 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.
@param alpha empirical weight that is applied to the VCI (set to 0.5). Please note that the weight of TCI is 1 - alpha.}
}
\value{
Spatially-distributed VHI values.
}
\description{
Vegetation Health Index (VHI; Kogan 1995)
}
