% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meteorological_drought_utils.R
\name{spi_agregate_daily}
\alias{spi_agregate_daily}
\title{Accumulations to be used by the daily_spi and daily_spei functions,}
\usage{
spi_agregate_daily(Prod_data, scale, trgt = NULL)
}
\arguments{
\item{Prod_data}{'SpatRaster' object that contains spatially-distributed monthly precipitation data that will be used to calculate the SPI. 
This 'SpatRaster' must include the time that corresponds to the dates of the respective layers. They can be set with the function time
of the terra package.}

\item{scale}{Integer value that represents the time scale at which the SPI will be computed.}
}
\value{
This function returns a 'SpatRaster' of the specific date (e.g., "%Y-02-28") according to the defined scale, defined as a 30-day 
  accumulation period (e.g., scale = 3, accumulates the specific date plus the previous 89 days).
}
\description{
Accumulations to be used by the daily_spi and daily_spei functions,
}
