% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydrological_drought.R
\name{ssi}
\alias{ssi}
\title{Wrapper function from the SCI package to calculate the Standardized Streamflow Index (SSI)}
\usage{
ssi(x, scale, ref_start = NULL, ref_end = NULL, distribution)
}
\arguments{
\item{x}{Zoo object with one or multiple monthly time series.}

\item{scale}{Integer value that represents the time scale at which the SSI will be computed.}

\item{ref_start}{optional value that represents the starting point of the reference period used for computing the index. 
The date should be introduced as '\%Y-\%m'. For example: "1989-02".
The default is NULL, which indicates that the first layer in the 'SpatRaster' will be used as starting point.}

\item{ref_end}{Optional value that represents the ending point of the reference period used for computing the index. 
The date should be introduced as '\%Y-\%m'. For example: "1989-02".
The default is NULL, which indicates that the last layer in the 'SpatRaster' will be used as ending point.}

\item{distribution}{Optional value indicating the name of the distribution function to be used for computing the SSI 
(one of 'log-Logistic', 'GEV', 'Gamma', and 'PearsonIII').
#'}
}
\value{
Standardised Streamflow Index.
}
\description{
Wrapper function from the SCI package to calculate the Standardized Streamflow Index (SSI)
}
