% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPI_SPEI_at_specific_day.R
\name{write_parameters}
\alias{write_parameters}
\title{Export the params_list object to a specific directory}
\usage{
write_parameters(params_list, dir, folder_name = NULL, export = FALSE)
}
\arguments{
\item{params_list}{A list object that contains a 'SpatRaster' of 365 layers for each parameter of the selected distribution.}

\item{dir}{Path to the directory to save the data}

\item{folder_name}{Which is the name of the parent folder that will be created to store the results? Defaults to NULL (name of folder = Results)}

\item{export}{Logical. Should the object be exported. Defaults to FALSE, and therefore, if it is not actively changed by the user,
it will not write anything to the directory}
}
\value{
Writes the params_list object to a specific directory
}
\description{
Export the params_list object to a specific directory
}
