% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_annual_stack.R
\name{.get_names}
\alias{.get_names}
\title{Get layer names, extension, and dates}
\usage{
.get_names(rst_path, start_date, end_date)
}
\arguments{
\item{start_date}{Position where the dates of the raster files start. For example, start_date = 10 for 'ProductX_1989-02-28.tif' as the date start in the 10th character.}

\item{end_date}{Position where the dates of the raster files end. For example, end_date = 19 for 'ProductX_1989-02-28.tif', and end_date = 16 for 'ProductX_1989-02.nc'.}
}
\value{
this function retrieves the file names (including the path), their corresponding dates and file extension.
}
\description{
Get layer names, extension, and dates
}
