% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trend_analyses.R
\name{.Senslope}
\alias{.Senslope}
\title{Utils function that calculates the Sen's slope}
\usage{
.Senslope(x, conflevel)
}
\arguments{
\item{x}{numerical vector.}

\item{conflevel}{Level of significance. Set to 0.95 by default.}
}
\description{
Utils function that calculates the Sen's slope
}
