% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Agricultural_drought_utils.R
\name{.vci_period}
\alias{.vci_period}
\title{Utils function to calculate the Vegetation Condition Index per period}
\usage{
.vci_period(x, dates, ...)
}
\arguments{
\item{x}{Numerical vector.}

\item{dates}{Vector of dates passed from the parent function.}
}
\value{
Numerical vector with the corresponding to the VCI
}
\description{
Utils function to calculate the Vegetation Condition Index per period
}
