% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Climdex.R
\name{R99p}
\alias{R99p}
\title{Annual total PRCP when RR > 99th percentile}
\usage{
R99p(
  rst.path,
  vct = NULL,
  temporal.scale = c("total", "annual"),
  start.date,
  end.date,
  date.fmt = "\%Y-\%m-\%d",
  pattern = NULL
)
}
\arguments{
\item{rst.path}{File path to daily raster files for the period of analysis. 
These files should include the date in any format.}

\item{vct}{Vector file of the study area (Optional). It will be used to crop 
the spatial extent of the raster files if required. This parameter is set 
to 'NULL'}

\item{temporal.scale}{either 'total' to use all the period of record, or
'annual' to compute the index annually (From Jan to Dec)}

\item{start.date}{Position where the dates of the raster files start}

\item{end.date}{Position where the dates of the raster files end}

\item{date.fmt}{Format of the dates included in the file names (default = 
"%Y-%m-%d")}

\item{pattern}{Set to NULL. Is there a specific pattern to list the 
raster files?}
}
\value{

}
\description{
Annual total PRCP when RR > 99th percentile
}
